Practice sketches for p5.js
sketch 1 is a recreated mathematical concept of collatz conjecture.

Sketch 2 is a recreated version of one of my sketches on processing. Helped me understand the differences between the two platforms. 

Sketch 3 was made use an existing example from the p5 library and changing it's parameters. The aim was to understand the code for the sketch and then be able to recreate it.

Sketch 4 was started from scratch.